Move heart and AuthType out of http
This file is going to get blasted in favor of Express.
