+++ test/update.test.ts
@@ -3,12 +3,11 @@ import * as fs from "fs-extra"
 import * as http from "http"
 import * as path from "path"
 import { LatestResponse, UpdateHttpProvider } from "../src/node/app/update"
-import { AuthType } from "../src/node/http"
+import { AuthType } from "../src/node/cli"
 import { SettingsProvider, UpdateSettings } from "../src/node/settings"
 import { tmpdir } from "../src/node/util"
 
-describe("update", () => {
-  return
+describe.skip("update", () => {
   let version = "1.0.0"
   let spy: string[] = []
   const server = http.createServer((request: http.IncomingMessage, response: http.ServerResponse) => {
